Move two tests
